This makes the manifest track master for everything...
This isn't always a good idea with OE.  Master's not one that is for the
faint of heart.  Oftentimes you're having to "fix" a lot of brokenness
in the tip of master with OE- because that's the wrecking ball branch
for the latest of things.  It's truly the bleeding edge and there are
few assurances of a clean build for any one given update to master.

Having said this, I'm working on some stuff intended to be eventually
upstreamed into git.yoctoproject.org, so I need a manifest that tracks
master for anything and everything.  The master branch for my manifest
for things will always track the latest or next to latest stable,
**RELEASED** branch from Poky/OE so that there's something KNOWN to
actually work.  All metadata in meta-pha will be validated to work
against that manifest and this one before commit- unless otherwise
noted because of a need to properly stow something in the GitHub.