Stripped out meta-rpi, meta-qt5.
There's too much deviation from what I need at this time with the
meta-rpi layer for baseline projects.  I need to understand what
he's crippled out for his own projects (sigh...if it's on GitHub,
it might behoove you to have some baselines that don't do things
like cripple out the BT on a Pi3, etc. along with a branch that
DOES do this...)

So, out it comes and I grabbed the upstream on master on meta-
raspberrypi and it's building 4.4.7 the "right" way (i.e. Stock
build with everything turned ON so the WiFi, etc. work)

Change-Id: I7f64309ddb4c4f66d70bca336b543f8a61c421e9