Removed a few layers...
OpenWRT's wonky on so many lavels...will revisit the metadata layers
there shortly with an eye on bringing key pieces in from it to my
PHA layer as we only really want things like the network management
daemon, etc. from it.  By and of itself, there's some wonky stuff
in OpwnWRT itself (which is part of why the current "maintained"
meta-openwrt is not robust enough to build right...sigh...)

Change-Id: I91b7ee68d0c6c128f7d20aae3216b229bf95e69d