Moved back from master to rocko for the userspace pieces...
The main body of things should be on _rocko_ which is more stable
in nature for userland builds, etc. than master (since master's
the OE equivalent of Debian's _*SID*_ which is a wrecking ball...)

This should be clean for most purposes and can be slowly expanded
upon with OpenWRT-ish metadata slowly being added to meta-pha, etc.
becase the meta-openwrt layer didn't stick with building the core
and then extending it...it throws the kitchen sink in and tries
to do the full monty for OpenWRT in one pass, which is...sadly...
broken right now.

Change-Id: Ic3ece8fff7921cbf1d9f05e2721d20d5bfdd58f1