Moved back to master for everything...
While it looked like it was going to build right, some things didn't
build right on a fresh build of the kernel (unacceptable...so...)
so back we go to unstable for everything for now.  It's not like
I've not dealt with THAT particular minefield before in OE- need
to support the new toys that I've got and that's the only real way
for right now, it seems.

Change-Id: I6947161ea214b03a76f0a78062de3f735935f57c