Removed golang support for now...
It does network-centric dependency checking, etc. which means without
network function, it breaks.  We'll include it on any forked off
versions of this manifest.  This is the core manifest for PHA Tech's and
correspondingly my projects, as well as any consumers of my handiwork
supporting FriendlyElec's and other vendor's boards that doesn't get
upstreamed/merged with other metadata layers for OE.  Fork it if we
need the golang support with the understanding of the issues there.

Change-Id: I6528772c5baaea17c65f730348f691f7cfb24b53