Removed golang support, moved meta-raspberrypi to Sumo
meta-raspberrypi's team released a Sumo version.  Rocko layer
and Sumo layer here incompatible.  Need to have Sumo because they
were using something that was deprecated away in their recipe for
the kernel there.

meta-golang was in the manifest to provide library packaging
support for the new go.bbclass functionality in from Rocko
and more gelled in Sumo, using the rocko version of this
layer to do it since it was compatible with both revisions
of core.  Problem is?  Go support cowpiles to a segfaulting
executable right now on at least aarch64.  Removing until
I can figure out what the heck gives there.

Change-Id: Ieb9a0848494540462728e17f744881c6cc2d8168