Revert "Added a couple of metadata layers to projects"
This reverts commit dcd376395659d1620bd4e3363d808ec3f176403b.  The
layers in question were...heh...embarassing.  The Haskell one was
broken in several ways that if I'm adding Haskell, we'll cherry-
pick things into meta-pha instead.  The meta-libretro is as broken
for cross-compilation as RetroArch and libretro is.   It desperately
needs to be re-worked to be using SOUND design practices...which
may never come to pass as it's not "cool" to make that and this
bunch is all about "cool".  I might revisit this eventually as
I'd love to have a universal engine to make "Lakka" anywhere
where we have GLES/Vulkan and ARM/MIPS/RISC-V/etc  That's just
not going to happen with this stuff right now, though.

Change-Id: I603f7ce484f24a15adf37a06dc5fa227a234f8e0